requirements.txt
fastapi
uvicorn
cryptography
PyJWT
argon2-cffi
pytest
pytest-asyncio
httpx